Refactor: PlanCardList에서 map으로 돌리는 컴포넌트로 PlanCard.tsx 사용
