--- src/components/main/planList/planCardList/PlanCard.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function PlanCard() {
-  return (
-    <div>
-      <div></div>
-    </div>
-  );
-}
